fix bug where done/undone ticjets wouldn't show right
Now the tickets I render are based on a new state of the main App component called ticketsToRender. This is done so that there is a separation between tickets which are fetched from the database and the ones that I want to show on screen (depending on what user chooses).

Also,  the stuff that I added to toggleTicketDone is for updating my tickets when I toggle between ticket's done state.